Revert "get rid of symlinks from docker-in-docker jank"
This reverts commit e8283e5d3d56e6bb4b0a12214458229602a28f7e.

+++ src/builder.py
@@ -1,4 +1,5 @@
 import asyncio
+import os
 import subprocess
 import sys
 import time
@@ -47,8 +48,6 @@ def build(job: ActionResult):
                 stdout=subprocess.PIPE,
                 stderr=subprocess.PIPE,
             )
-            job.log(output.stdout)
-            job.log(output.stderr)
         except subprocess.CalledProcessError as e:
             job.on_error(
                 e, f"[BUILD] Failed to build commit {job.commit.hash}! No commit found."
@@ -62,11 +61,10 @@ def build(job: ActionResult):
         try:
             output = subprocess.run(
                 "cd ectf-design-repo &&"
-                "rm -rf ~/mounts/secrets/* &&"
-                "mkdir -p ~/mounts/secrets &&"
+                "rm -rf secrets/* &&"
+                "mkdir -p secrets &&"
                 ". ./.venv/bin/activate &&"
-                "pip install -e ./ectf26_design &&"
-                "python -m gen_secrets ~/mounts/secrets/global.secrets 1 2 3 4",
+                "uv run secrets ./secrets/global.secrets 1 2 3 4",
                 shell=True,
                 check=True,
                 stdout=subprocess.PIPE,
@@ -86,28 +84,37 @@ def build(job: ActionResult):
         job.log("[BUILD] Building firmware...")
         # build firmware
         try:
-            # docker-in-docker jank
-            # build_server_build_out is volume mounted to ~/mounts/build_out
-            # build_server_secrets is volume mounted to ~/mounts/secrets
-            # build_server_firmware is volume mounted to ~/mounts/firmware
-            output = subprocess.run(
-                "cd ectf-design-repo &&"
-                "rm -rf build_out/* ~/mounts/firmware/* &&"
-                "(docker build -t build-hsm ./firmware &&"
-                "cp -r ./firmware/* ~/mounts/firmware &&"
-                "docker run --rm "
-                "-v build_server_firmware:/hsm "
-                "-v build_server_secrets:/secrets "
-                "-v build_server_build_out:/out "
-                "-e HSM_PIN='1a2b3c' "
-                "-e PERMISSIONS='1234=R--:4321=RWC' build-hsm) && "
-                '[ -n "$(ls -A ~/mounts/build_out 2>/dev/null)" ]',
-                shell=True,
-                check=True,
-                timeout=60 * 10,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-            )
+            if os.getenv("DOCKER"):
+                # docker-in-docker jank
+                # build_server_build_out is volume mounted to ~/mounts/build_out which is symlinked to ~/src/ectf-design-repo/build_out
+                # build_server_secrets is volume mounted to ~/mounts/secrets which is symlinked to ~/src/ectf-design-repo/secrets
+                # build_server_firmware is volume mounted to ~/mounts/firmware which is symlinked to ~/src/ectf-design-repo/firmware
+                output = subprocess.run(
+                    "cd ectf-design-repo &&"
+                    "rm -rf build_out/* ~/mounts/firmware/* &&"
+                    "(docker build -t build-hsm ./firmware &&"
+                    "cp -r ./firmware/* ~/mounts/firmware &&"
+                    "docker run --rm -v build_server_firmware:/hsm "
+                    "-v build_server_secrets:/secrets "
+                    "-v build_server_build_out:/out -e HSM_PIN='1a2b3c' "
+                    "-e PERMISSIONS='1234=R--:4321=RWC' build-hsm) && "
+                    '[ -n "$(ls -A build_out 2>/dev/null)" ]',
+                    shell=True,
+                    check=True,
+                    timeout=60 * 10,
+                    stdout=subprocess.PIPE,
+                    stderr=subprocess.PIPE,
+                )
+            else:
+                output = subprocess.run(
+                    "cd ectf-design-repo && ./build.sh && "
+                    '[ -n "$(ls -A build_out 2>/dev/null)" ]',
+                    shell=True,
+                    check=True,
+                    timeout=60 * 10,
+                    stdout=subprocess.PIPE,
+                    stderr=subprocess.PIPE,
+                )
             job.log(output.stdout)
             job.log(output.stderr)
         except subprocess.SubprocessError as e:
@@ -236,6 +243,14 @@ def init_build_queue():
             check=True,
             stdout=subprocess.PIPE,
             stderr=subprocess.PIPE,
+        )
+    if os.getenv("DOCKER"):  # setup for docker-in-docker jank
+        subprocess.run(
+            "rm -rf ./ectf-design-repo/secrets ./ectf-design-repo/build_out;"
+            "ln -s ~/mounts/secrets ./ectf-design-repo/secrets;"
+            "ln -s ~/mounts/build_out ./ectf-design-repo/build_out;",
+            shell=True,
+            check=True,
         )
 
     # create venv
